drftutorial/requirements.txt
django
djangorestframework
django-extensions
djangorestframework-simplejwt